# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AdminSetFriendBanRequest:
    properties:
      banned_until:
        type: string
      reason:
        type: string
    required:
    - banned_until
    - reason
    type: object
...
  model.FileListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/model.FileResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  model.FileResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      mime_type:
        type: string
      original_name:
        type: string
      size:
        type: integer
      storage_name:
        type: string
      storage_type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
...
  response.ResponseData:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error: {}
      message:
        example: 成功
        type: string
      timestamp:
        example: 1640995200
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 这是一个用户注册和认证系统的后端API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
  /admin/files:
    get:
      description: 分页筛选所有文件（公开与私有）
      parameters:
      - description: 文件分类
        in: query
        name: category
        type: string
      - description: 存储类型
        in: query
        name: storage_type
        type: string
      - description: 存储名称
        in: query
        name: storage_name
        type: string
      - description: 是否公开
        in: query
        name: is_public
        type: boolean
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员获取文件列表
      tags:
      - admin-files
  /admin/files/{id}:
    delete:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员删除文件
      tags:
      - admin-files
    get:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员获取文件详情
      tags:
      - admin-files
    put:
      consumes:
      - application/json
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.FileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员更新文件
      tags:
      - admin-files
  /admin/files/public:
    get:
      description: 分页筛选公开文件
      parameters:
      - description: 文件分类
        in: query
        name: category
        type: string
      - description: 存储类型
        in: query
        name: storage_type
        type: string
      - description: 存储名称
        in: query
        name: storage_name
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员获取公开文件列表
      tags:
      - admin-files
  /admin/logs:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 管理员用户名
        in: query
        name: admin_username
        type: string
      - description: 操作类型
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 查询管理员行为日志
      tags:
      - admin-logs
    post:
      consumes:
      - application/json
      parameters:
      - description: 日志创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AdminLogCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 创建管理员行为日志
      tags:
      - admin-logs
...
  /admin/users:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 管理员获取用户列表
      tags:
      - admin-users
  /admin/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员删除用户
      tags:
      - admin-users
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
      summary: 管理员获取用户详情
      tags:
      - admin-users
  /admin/users/{id}/action-logs:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 管理员查询用户行为日志
      tags:
      - admin-users
  /admin/users/{id}/friend-ban:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员解除用户好友功能封禁
      tags:
      - admin-users
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员查询用户好友功能封禁状态
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 封禁请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AdminSetFriendBanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员设置用户好友功能封禁
      tags:
      - admin-users
  /admin/users/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 密码更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AdminUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员更新用户密码
      tags:
      - admin-users
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 管理员更新用户状态
      tags:
      - admin-users
  /files/{id}:
    delete:
      description: 删除指定的文件（仅文件所有者可操作）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ResponseData'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除文件
      tags:
      - files
    get:
      description: 根据文件ID获取文件详细信息
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取文件详情
      tags:
      - files
    put:
      consumes:
      - application/json
      description: 更新文件的分类、描述等信息（仅文件所有者可操作）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.FileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ResponseData'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新文件信息
      tags:
      - files
  /files/my:
    get:
      description: 获取当前登录用户的文件列表，支持分页和筛选
      parameters:
      - description: 文件分类筛选
        in: query
        name: category
        type: string
      - description: 存储类型筛选
        in: query
        name: storage_type
        type: string
      - description: 存储名称筛选
        in: query
        name: storage_name
        type: string
      - description: 是否公开筛选
        in: query
        name: is_public
        type: boolean
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileListResponse'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户的文件列表
      tags:
      - files
  /files/public:
    get:
      description: 获取所有公开访问的文件列表，支持分页和筛选
      parameters:
      - description: 文件分类筛选
        in: query
        name: category
        type: string
      - description: 存储类型筛选
        in: query
        name: storage_type
        type: string
      - description: 存储名称筛选
        in: query
        name: storage_name
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileListResponse'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取公开文件列表
      tags:
      - files
  /files/storages:
    get:
      description: 获取系统可用的存储配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.StorageInfoResponse'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取存储信息
      tags:
      - files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件到指定的存储位置
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      - description: 存储名称（可选，默认使用系统默认存储）
        in: formData
        name: storage_name
        type: string
      - description: 文件分类
        in: formData
        name: category
        type: string
      - description: 文件描述
        in: formData
        name: description
        type: string
      - description: 是否公开访问
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.FileResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ResponseData'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 上传单个文件
      tags:
      - files
  /files/upload-multiple:
    post:
      consumes:
      - multipart/form-data
      description: 批量上传多个文件到指定的存储位置
      parameters:
      - collectionFormat: csv
        description: 要上传的文件列表
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      - description: 存储名称（可选，默认使用系统默认存储）
        in: formData
        name: storage_name
        type: string
      - description: 文件分类
        in: formData
        name: category
        type: string
      - description: 文件描述
        in: formData
        name: description
        type: string
      - description: 是否公开访问
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FileResponse'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ResponseData'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 上传多个文件
      tags:
      - files
  /friends/{friend_id}:
    delete:
      parameters:
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 删除好友
      tags:
      - 好友
  /friends/blocks:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 黑名单列表
      tags:
      - 好友
  /friends/blocks/{user_id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 取消拉黑
      tags:
      - 好友
    post:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 拉黑用户
      tags:
      - 好友
  /friends/list:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取好友列表
      tags:
      - 好友
  /friends/remarks/{friend_id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 更新好友备注
      tags:
      - 好友
  /friends/requests:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 发起好友请求
      tags:
      - 好友
  /friends/requests/{id}:
    delete:
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 撤回好友请求
      tags:
      - 好友
  /friends/requests/{id}/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 接受好友请求
      tags:
      - 好友
  /friends/requests/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 拒绝好友请求
      tags:
      - 好友
  /friends/requests/incoming:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 入站好友请求列表
      tags:
      - 好友
  /friends/requests/outgoing:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 出站好友请求列表
      tags:
      - 好友
  /users/{id}:
    get:
      consumes:
      - application/json
      description: 通过用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 根据ID获取用户信息
      tags:
      - 用户管理
  /users/activate:
    post:
      consumes:
      - application/json
      description: 使用验证码激活非活跃账户
      parameters:
      - description: 账户激活请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ActivateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 账户激活成功
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: 请求参数错误或验证码错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 激活账户
      tags:
      - 用户管理
  /users/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，成功后返回包含Access Token、Refresh Token和用户信息的对象
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResponse'
              type: object
        "400":
          description: 请求参数错误或设备验证码相关错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "403":
          description: 账户已被封禁或未激活
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 用户登录
      tags:
      - 用户管理
  /users/logout:
    post:
      consumes:
      - application/json
      description: 登出用户并撤销所有Token（Access Token和Refresh Token）。Access Token将被加入黑名单，立即失效。
      parameters:
      - description: 登出请求（需要提供access_token和refresh_token）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ResponseData'
        "401":
          description: Token无效或两个Token不属于同一用户
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 用户登出
      tags:
      - 用户管理
...
  /ws/chat:
    get:
      description: '通过 WebSocket 建立一对一聊天连接。鉴权方式：优先读取 HTTP Header `Authorization: Bearer
        <access_token>`；浏览器无法自定义 Header 时可使用 query 参数 `?token=<access_token>`。'
      parameters:
      - description: Access Token（可选，浏览器场景使用）
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      summary: WebSocket 聊天连接
      tags:
      - chat
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"